Add new 'wal' and 'memory' log stores
This commit includes two new log store formats, 'wal' and 'memory'.

The 'wal' store, which is now the default, is a write-ahead log format that
is designed specifically for raft-type programs. It's about 25% faster than
existing disk-based formats, such as 'leveldb'.

The implementation is at https://github.com/tidwall/wal.

The 'memory' store is a pure in-memory log store. The memory store will not
persist to disk. Use with caution.

A backend log format can be selected using the startup flag
`--backend wal|memory|leveldb|bolt`.

Also included with this commit is a new stable store 'json', which is now the
primary stable store format.